// apps/web/src/app/features/inventory/modules/drug-generics/types/drug-generics.types.ts
export interface DrugGeneric {
  id: number;
  working_code: string;
  generic_name: string;
  dosage_form?: string | null;
  strength_unit?: string | null;
  dosage_form_id?: number | null;
  strength_unit_id?: number | null;
  strength_value?: number | null;
  is_active?: boolean | null;
  created_at?: string | null;
  updated_at?: string | null;
}